padel/requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
cssselect==1.2.0
Django==4.1.3
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-filter==22.1
django-seed
psycopg2
django-redis
django-htmx==1.13.0
django_debug_toolbar==3.8.1
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lxml==4.9.1
mypy-extensions==0.4.3
packaging==21.3
parsel==1.7.0
pathspec==0.10.2
platformdirs==2.5.4
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
w3lib==2.1.0
whitenoise==5.3.0
zope.interface==5.5.2
gunicorn
celery==5.1.2
redis==3.5.3